% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1.R
\name{d1_01}
\alias{d1_01}
\title{The L1-distance}
\usage{
d1_01(D)
}
\arguments{
\item{D}{A design}
}
\value{
The return value is the L1-distance of design D.
}
\description{
Calculate the L1-distance of a design when the design is scaled into [0,1]^n
}
\examples{
library(MaximinDesign)
D <- EquidistantD(3)
d1_01(D)
}
