% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1.R
\name{d1_eff}
\alias{d1_eff}
\title{The L1-distance efficiency}
\usage{
d1_eff(D)
}
\arguments{
\item{D}{A U-type design}
}
\value{
The return value is the L1-distance efficiency of design D.
}
\description{
Calculate the L1-distance efficiency of a U-type design D whose levels are from 0,1, ..., s-1, where D is said to be U-type if the number of every level appears equally often in every column.
}
\examples{
library(MaximinDesign)
D <- EquidistantD(3)
d1_eff(D)
}
