% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2.R
\name{d2}
\alias{d2}
\title{The L2-distance}
\usage{
d2(D)
}
\arguments{
\item{D}{A design}
}
\value{
The return value is the L2-distance of a design.
}
\description{
Calculate the L2-distance of a design when the design whose levels are from 0,1, ..., s-1
}
\examples{
library(MaximinDesign)
D <- EquidistantD(3)
d2(D)
}
\references{
Wenlong Li, Min-Qian Liu and Boxin Tang (2021). A method of constructing maximin distance designs.   \emph{Biometrika}, published online. <doi:10.1093/biomet/asaa089>
}
