% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2.R
\name{d2_01}
\alias{d2_01}
\title{The L2-distance}
\usage{
d2_01(D)
}
\arguments{
\item{D}{A design}
}
\value{
The return value is the L2-distance of design D.
}
\description{
Calculate the L2-distance of a design when the design is scaled into [0,1]^n
}
\examples{
library(MaximinDesign)
D <- EquidistantD(3)
d2_01(D)
}
