% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2.R
\name{d2_eff}
\alias{d2_eff}
\title{The L2-distance efficiency}
\usage{
d2_eff(D)
}
\arguments{
\item{D}{A U-type design}
}
\value{
The return value is the L2-distance efficiency of design D.
}
\description{
Calculate the L2-distance efficiency of a U-type design D whose levels are from 0,1, ..., s-1, where D is said to be U-type if the number of every level appears equally often in every column.
}
\examples{
library(MaximinDesign)
D <- EquidistantD(3)
d2_eff(D)
}
