% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquidistantD.R
\name{EquidistantD}
\alias{EquidistantD}
\title{Equi-distant design}
\usage{
EquidistantD(s)
}
\arguments{
\item{s}{The run size of design B, if 2s + 1 is a prime, e.g.,  s=3,5,8,9,11,....}
}
\value{
The return value is equi-distant LH(s,s).
}
\description{
\code{EquidistantD} constructs an equi-distant design, and then provides a choice of design B.
}
\details{
Under the L1-distance, Wang et al. (2018,Theorem4) showed that an equi-distant LH(s,s), a Latin hypercube of s runs for s factors, can be constructed if 2s + 1 is a prime.
}
\examples{
# EquidistantD(s) for s=3,5,8,9,11,...
B <- EquidistantD(3)
B <- EquidistantD(5)
}
\references{
Wenlong Li, Min-Qian Liu and Boxin Tang (2021). A method of constructing maximin distance designs.   \emph{Biometrika}, published online. <doi:10.1093/biomet/asaa089>
}
