% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximinD.R
\name{MaximinD}
\alias{MaximinD}
\title{Maximin distance design}
\usage{
MaximinD(A, B)
}
\arguments{
\item{A}{The initial U-type design design U(N,s^n1), a design with N runs, n1 factors, and s levels 0,1,...,s - 1.}

\item{B}{The initial U-type design design U(s,q^n2), a design with s runs, n2 factors, and s levels 0,1,...,q - 1.}
}
\value{
The return value is a U-type design U(N,q^(n1*n2)), a design with N runs, n1*n2 factors, and q levels from 0,1,...,q - 1.
}
\description{
\code{MaximinD} is a method of constructing maximin distance designs
}
\details{
Specifically, we construct a U-type design D by replacing the \emph{u}th level of A by the  (\emph{u} + 1)th row of B for \emph{u} = 0,1,...,s - 1,  where D is said to be U-type if the number of every level appears equally often in every column.
}
\examples{
library(MaximinDesign)

A <- SaturatedOA(s = 3, k = 2)
B <- EquidistantD(3)
D <- MaximinD(A, B)

# To obtain the results in Table 5 in Example 2, we give the following code.
library(MaximinDesign)
s <- 5
k <- 3
S <- SaturatedOA(s, k); n1 <- ncol(S)
B <- EquidistantD(s)
SequentialDel <-  function(n){D <- MaximinD(S[,(n1-n+1):n1], B); r <- d1_eff(D);r}
col_lab <- 31:19
cbind(s*col_lab, apply(as.matrix(col_lab), 1, SequentialDel))



}
\references{
Wenlong Li, Min-Qian Liu and Boxin Tang (2021). A method of constructing maximin distance designs.   \emph{Biometrika}, published online. <doi:10.1093/biomet/asaa089>
}
