% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prop2.R
\name{Prop2}
\alias{Prop2}
\title{Designs from Proposition 2}
\usage{
Prop2(s, k)
}
\arguments{
\item{s}{The number of levels}

\item{k}{As N = s^k is the run size of design D, k decides N.}
}
\value{
The return value is a U(N,s^n), design D
}
\description{
If s is a prime, \code{Prop2} is to construct a U(N,s^n), design D, where N = s^k and n = s^k - 1 in Proposition 2, as shown in Table 2.
For example, s=5,7,11,13,17,19,...
}
\examples{
# To construct designs in Proposition 2, as shown in Table 2, we gives two examples.
library(MaximinDesign)
D <- Prop2(s = 5, k = 2); # d1_eff(D)
D <- Prop2(s = 7, k = 2); d1_eff(D)
}
\references{
Wenlong Li, Min-Qian Liu and Boxin Tang (2021). A method of constructing maximin distance designs.   \emph{Biometrika}, published online. <doi:10.1093/biomet/asaa089>
}
