% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prop3.R
\name{Prop3}
\alias{Prop3}
\title{Designs from Proposition 3}
\usage{
Prop3(s, k1)
}
\arguments{
\item{s}{The number of levels}

\item{k1}{As N = s^k1 is the run size of design D, k1 decides N.}
}
\value{
The return value is a U(N,s^n), design D
}
\description{
Let s = 2^k2 be a power of 2, \code{Prop3} is to construct a U(N,s^n), design D, where N = s^k1 and n = 2^(k2-1)*(N-1)/(2^k2-1) in Proposition 3.
}
\examples{
# To construct designs in Proposition 3, we gives three examples.
library(MaximinDesign)
D <- Prop3(s = 4, k1 = 2); d2_eff(D)
D <- Prop3(s = 8, k1 = 2); d2_eff(D)
D <- Prop3(s = 16, k1 = 2); d2_eff(D)

}
\references{
Wenlong Li, Min-Qian Liu and Boxin Tang (2021). A method of constructing maximin distance designs.   \emph{Biometrika}, published online. <doi:10.1093/biomet/asaa089>
}
